+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - add experimental [Time](https://clickhouse.com/docs/sql-reference/data-types/time) and [Time64](https://clickhouse.com/docs/sql-reference/data-types/time64) types support https://github.com/plausible/ch/pull/260
+- add [Variant](https://clickhouse.com/docs/sql-reference/data-types/variant) type support https://github.com/plausible/ch/pull/263
 
 ## 0.4.1 (2025-07-07)